requirements.txt
Flask==3.1.0
google-generativeai==0.8.5
requests==2.32.3
python-socketio[client]==5.13.0
python-dotenv==1.1.0
Pillow==11.2.1
instaloader==4.14.1
faster-whisper==1.1.1